deleteTemplate
